requirements.txt
pandas
faker
google-generativeai
